requirements.txt
fredapi
python-dotenv
xgboost
sklearn